+++ example/chinese_lunar_calendar_example.dart
@@ -1,8 +1,8 @@
 import 'package:chinese_lunar_calendar/chinese_lunar_calendar.dart';
 
 void main() {
-  final lunarCalendar =
-      LunarCalendar(utcDateTime: DateTime(2023, 6, 5, 10, 30).toUtc());
+  final lunarCalendar = LunarCalendar(
+      utcDateTime: DateTime(2023, 6, 5, 10, 30).toUtc(), isSimplifiedCN: true);
 
   print('中国标准时间：${lunarCalendar.cst}');
   print('本地时间：${lunarCalendar.localTime}');