+++ lib/chinese_lunar_calendar.dart
@@ -3,3 +3,4 @@ library chinese_lunar_calendar;
 export 'src/lunar_calendar/lunar_calendar.dart';
 export 'src/lunar_calendar/models/models.dart';
 export 'src/utils/utils.dart';
+export 'src/localization/localization.dart';